changed xrd parser folder
