+++ src/nomad_measurements/xrd/parser.py
